Corrected MultiKeyMap
Removed MultiKeyMap and replaced it with MultiKeyReversableMap which includes undefined and null checks on functions.

Added ReversableMap which is the same as MultiKey but works with single keys instead of multiple.

Updated Utilities to import the new files into webpack

+++ Code/src/Utilities/MultiKeyReversableMap.ts
@@ -1,14 +1,14 @@
 import { CompareTypes } from "./CompareTypes";
 
-/** Class that handles multiple keys to one value Map
+/** Class that handles multiple keys to value Map with reversable search 
  */
- export class MultiKeyMap<Key, Value> {
+ export class MultiKeyReversableMap<Key, Value> {
 
     //** Store all forward references to values to allow searching*/
-    private __map__: Map<Key, Value> = new Map()
+    private __map__: Map<Key, Value> = new Map();
 
     /** Store all inverse references to values to allow searching and Maps the values to void for O(1) checking*/
-    private __reverseMap__: Map<Value, Map<Key, void>> = new Map()
+    private __reverseMap__: Map<Value, Map<Key, void>> = new Map();
 
     /**
 	 * Create a map between multiple keys and a single value
@@ -38,13 +38,22 @@ import { CompareTypes } from "./CompareTypes";
 	 */
     public getValue(key: Key): Value | undefined {
         return this.__map__.get(key);
-    }
+    }    
+    
+    /**
+    * Test if a key exists for a given value
+    * @param  {Value} value - The value to check if a key exists
+    */
+   public hasValue(value: Value): boolean {
+        return (this.getKeys(value) != undefined);
+   }
 
     /**
 	 * Return Map of Value to Keys
 	 * @param  {Value} Value - The value to return keys Map for
 	 */
     public getKeys(value: Value): Map<Key, void> | undefined {
+        if (value === undefined) { console.error("Trying to get a MultiKeyMap's Keys map with an invalid value: ", value); return undefined; }
         return this.__reverseMap__.get(value);
     }
 
@@ -53,6 +62,7 @@ import { CompareTypes } from "./CompareTypes";
 	 * @param  {Value} Value - The value to return keys for
 	 */
     public getKeysArray(value: Value): Key[] {
+        if (value === undefined) { console.error("Trying to get a MultiKeyMap's Keys array with an invalid value: ", value); return []; }
         let keys = this.getKeys(value);
         return keys != undefined ? Array.from(keys.keys()) : [];
     }
@@ -71,6 +81,9 @@ import { CompareTypes } from "./CompareTypes";
 	 * @param  {Value} value - The value to link this key to
 	 */
     public setKey(key: Key, value: Value) {
+        if (key == null) { console.error("Trying to set a MultiKeyMap's Key with an invalid key: ", key); return false; }
+        if (value === undefined) { console.error("Trying to set a MultiKeyMap's Key with an invalid value: ", value); return false; }
+
         let keys: Map<Key, void> | undefined = this.getKeys(value);
         if (keys == undefined) {
             keys = new Map();
@@ -82,6 +95,8 @@ import { CompareTypes } from "./CompareTypes";
             keys.set(key);
             this.__reverseMap__.set(value, keys);
         }
+
+        return true;
     }
 
     /**
@@ -89,10 +104,13 @@ import { CompareTypes } from "./CompareTypes";
 	 * @param  {Key | Key[]} Keys - Any keys to add this value to
 	 * @param  {Value} value - The value to link these keys to
 	 */
-    public setKeys(keys: Key[], value: Value) {
+    public setKeys(keys: Key[], value: Value): boolean {
+        if (keys == null) { console.error("Trying to set a MultiKeyMap's Keys with an invalid keys: ", keys); return false; }
+        if (value === undefined) { console.error("Trying to set a MultiKeyMap's Keys with an invalid value: ", value); return false; }
         keys.forEach(element => {
             this.setKey(element, value);
         });
+        return true;
     }
 
     /**
@@ -102,7 +120,7 @@ import { CompareTypes } from "./CompareTypes";
 	 */
      public remapValue(value: Value, newValue: Value): Key[] {
         if (value == null) { console.error("Trying to remap a MultiKeyMap's value with an invalid value: ", value); return []; }
-        if (newValue == null) { console.error("Trying to remap a MultiKeyMap's value with an invalid newValue: ", newValue); return []; }
+        if (newValue === undefined) { console.error("Trying to remap a MultiKeyMap's value with an invalid newValue: ", newValue); return []; }
         if (CompareTypes.different(value, newValue)) { console.error("Trying to remap a MultiKeyMap's value with different parameter types: ", value, " ", newValue); return []; }
 
         let keys = this.getKeysArray(value);
@@ -121,6 +139,9 @@ import { CompareTypes } from "./CompareTypes";
 	 * @param  {Value} newValue - The new value to link the keys to
 	 */
      public remapValueFromKey(key: Key, newValue: Value): Key[] {
+        if (key == null) { console.error("Trying to remap a value from a key in a MultiKeyReversableMap wtih an invalid key: ", key); return []; }
+        if (newValue === undefined) { console.error("Trying to remap a value from a key in a MultiKeyReversableMap wtih an invalid newValue: ", newValue); return []; }
+
         let oldValue = this.getValue(key);
         if (oldValue == undefined) return [];
         return this.remapValue(oldValue, newValue);
@@ -131,12 +152,16 @@ import { CompareTypes } from "./CompareTypes";
 	 * @param  {Key} key - The key to remap
 	 * @param  {Value} newValue - The new value to link the key to
 	 */
-     public remapKey(key: Key, newValue: Value) {
+     public remapKey(key: Key, newValue: Value): boolean {
+        if (key == null) { console.error("Trying to remap a key in a MultiKeyReversableMap wtih an invalid key: ", key); return false; }
+        if (newValue == null) { console.error("Trying to remap a key in a MultiKeyReversableMap wtih an invalid newValue: ", newValue); return false; }
+
         let oldValue = this.getValue(key);
-        if (oldValue == undefined) { this.setKey(key, newValue); return; }
+        if (oldValue == undefined) { this.setKey(key, newValue); return true; }
         
         this.deleteKey(key, oldValue);
         this.setKey(key, newValue);
+        return true;
     }
 
     /**
@@ -144,15 +169,18 @@ import { CompareTypes } from "./CompareTypes";
 	 * @param  {Key} Key - The key to remove from this value
 	 * @param  {Value} value - The value that this key links to
 	 */
-    public deleteKey(key: Key, value: Value) {
+    public deleteKey(key: Key, value: Value): boolean {
+        if (key == null) { console.error("Trying to delete a key in a MultiKeyReversableMap wtih an invalid key: ", key); return false; }
+        if (value == null) { console.error("Trying to delete a key in a MultiKeyReversableMap wtih an invalid value: ", value); return false; }
         let otherKeys = this.__reverseMap__.get(value);
 
-        if (otherKeys == undefined) { this.__map__.delete(key); return; }
+        if (otherKeys == undefined) { this.__map__.delete(key); return true; }
         otherKeys.delete(key);
-        if (otherKeys.size == 0) { this.__map__.delete(key); this.__reverseMap__.delete(value); return; }
+        if (otherKeys.size == 0) { this.__map__.delete(key); this.__reverseMap__.delete(value); return true; }
         
         this.__reverseMap__.set(value, otherKeys);
         this.__map__.delete(key);
+        return true;
     }
 
     /**
@@ -160,21 +188,26 @@ import { CompareTypes } from "./CompareTypes";
 	 * @param  {Key} Keys - The keys to remove from this value
 	 * @param  {Value} value - The value that these keys link to
 	 */
-    public deleteKeys(keys: Key[], value: Value) {
+    public deleteKeys(keys: Key[], value: Value): boolean {
+        if (keys == null || keys.length) { console.error("Trying to delete a keys in a MultiKeyReversableMap wtih an invalid keys: ", keys); return false; }
+        if (value == null) { console.error("Trying to delete a keys in a MultiKeyReversableMap wtih an invalid value: ", value); return false; }
         keys.forEach(element => {
             this.deleteKey(element, value);
         });
+        return true;
     }
 
     /**
 	 * Remove all references to a value
 	 * @param  {Value} value - The value to remove
 	 */
-    public deleteValue(value: Value) {
+    public deleteValue(value: Value): boolean {
+        if (value == null) { console.error("Trying to delete a value in a MultiKeyReversableMap wtih an invalid value: ", value); return false; }
         this.getKeysArray(value).forEach(element => {
             this.__map__.delete(element);
         });
         this.__reverseMap__.delete(value);
+        return true;
     }
 
     /**
